+++ app/src/main/java/com/example/seoulgonggong/data/datasource/GeoDataSource.kt
@@ -1,0 +1,8 @@
+package com.example.seoulgonggong.data.datasource
+
+interface GeoDataSource {
+    suspend fun getAddressByPosition(
+        latitude: Double,
+        longitude: Double,
+    ): String
+}
